Studio/requirements.txt
# OmniRoam Interactive System - Backend Dependencies

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Data validation
pydantic>=2.0.0

# Image/Video processing
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
imageio>=2.30.0
imageio-ffmpeg>=0.4.8

# PyTorch (install separately based on CUDA version)
# torch>=2.0.0
# torchvision>=0.15.0

# GPU monitoring (for accurate VRAM usage)
nvidia-ml-py>=12.0.0

# Async support
aiofiles>=23.0.0


